requirements.txt
# Core dependencies
requests==2.32.3
beautifulsoup4==4.13.3
pandas==2.2.3
tqdm==4.67.1
urllib3==2.3.0
pyyaml==6.0.2
spacy==3.8.4
tiktoken==0.8.0
# playwright>=1.49.0
python-dotenv==1.0.1
# ollama==0.4.7
google-generativeai
markdown
# pinecone-client==3.0.2
# langchain-ollama>=0.2.5
langchain>=0.1.0
chromadb>=0.4.24
aiohttp>=3.11.13
sentence-transformers>=2.2.2
langchain-community>=0.0.13
uuid>=0.0.0
openai==1.66.3
httpx==0.27.2
streamlit>=1.32.0

# Machine learning dependencies
numpy==1.26.4
transformers==4.37.0
datasets==2.14.5
torch==2.2.2
tensorboard==2.14.0
accelerate==0.21.0
scipy==1.13.0
bitsandbytes>=0.41.1